@@ -1,7 +1,9 @@
 ﻿<SectionContent SectionName="Header">
     <UserHeaderNav NavClass="breadcrumbs" NoUserLogin="true">
-        <NavLink href="" Match="NavLinkMatch.All">⌂</NavLink>
-        <NavLink>conesoft</NavLink> ▸ <NavLink href="" Match="NavLinkMatch.All">user access</NavLink>
+        <HomeButtonContent>⌂</HomeButtonContent>
+        <NavContent>
+            <NavLink>conesoft</NavLink> ▸ <NavLink href="" Match="NavLinkMatch.All">user access</NavLink>
+        </NavContent>
     </UserHeaderNav>
 </SectionContent>
 
