@@ -1,7 +1,9 @@
 ﻿<SectionContent SectionName="Header">
     <UserHeaderNav NavClass="breadcrumbs">
-        <NavLink href="" Match="NavLinkMatch.All">⌂</NavLink>
-        <NavLink>conesoft</NavLink> ▸ <NavLink>settings</NavLink> ▸ <NavLink href="" Match="NavLinkMatch.All"><Username /></NavLink>
+        <HomeButtonContent>⌂</HomeButtonContent>
+        <NavContent>
+            <NavLink>conesoft</NavLink> ▸ <NavLink>settings</NavLink> ▸ <NavLink href="" Match="NavLinkMatch.All"><Username /></NavLink>
+        </NavContent>
     </UserHeaderNav>
 </SectionContent>
 
