@@ -3,7 +3,7 @@
 
 <PageTitle>Conesoft</PageTitle>
 <HeadContent>
-    <HtmlHead />
+    <HtmlHead Cdn="new" />
 
     <meta name="description" content="Conesoft" />
     <meta property="og:type" content="website" />
