@@ -4,13 +4,13 @@ namespace Conesoft.Website.Utilities;
 
 public static class DirectoryWatcher
 {
-    public static async Task WatchLive(this Conesoft.Files.Directory directory, Action<Conesoft.Files.File[]> callback)
+    public static async Task WatchLive(this Conesoft.Files.Directory directory, Action<IEnumerable<Conesoft.Files.File>> callback)
     {
         await Task.Run(async () =>
         {
-            await foreach (var files in directory.Live(allDirectories: true).Changes())
+            await foreach (var entries in directory.Changes(allDirectories: true))
             {
-                callback(files.All);
+                callback(entries.All.Files());
             }
         });
     }
